# apps/core-stub/openapi.yaml
openapi: 3.1.0
info:
  title: plataforma-core-ohs (stub)
  version: 1.0.0
  description: |
    Mock del sistema core de la aseguradora. Expone catálogos maestros, datos de
    suscriptores, agentes y configuración de pólizas. En producción este servicio
    sería el ERP/core de la compañía; en el reto se implementa como stub con fixtures.
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo local
paths:
...
  /v1/zip-codes/{zipCode}:
    get:
      summary: Obtener datos de zona por código postal
      operationId: getZipCode
      tags: [ZipCodes]
      parameters:
        - name: zipCode
          in: path
          required: true
          schema:
            type: string
            example: '170101'
      responses:
        '200':
          description: Datos de zona
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZipCodeZone'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/zip-codes/validate:
    post:
      summary: Validar si un código postal es reconocido
      operationId: validateZipCode
      tags: [ZipCodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codigoPostal]
              properties:
                codigoPostal:
                  type: string
      responses:
        '200':
          description: Resultado de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  codigoPostal:
                    type: string
                  valid:
                    type: boolean
...
components:
  schemas:
...
    ZipCodeZone:
      type: object
      properties:
        codigoPostal:
          type: string
        municipio:
          type: string
        provincia:
          type: string
        pais:
          type: string
        zonaRiesgoIncendio:
          type: string
        zonaRiesgoSismo:
          type: string
        zonaFhm:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...